Created the page for Teams, Matches and Players, created some forms in order to edit and create new teams, matches and players
Created button for delete.

Added nav bar.
Implemented the leagueServices
created the endpoints and tested the comunication
Changed the URL because the port was changed.
created and connected the endpoint with the frontend

There are a few issues that cannot fix yet because the error is on the back end, I continue editing code in the back end in order to maintain the good practices and keep this running.